+++ ImageOptimizer.php
@@ -2,4 +2,5 @@
 
 //FileCompiler=0
 
+// Image optimizer: https://github.com/psliwa/image-optimizer
 require_once __DIR__ . '/image-optimizer/vendor/autoload.php';